+++ CHANGE_LOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 2016-09-14 Release 2.0.10:
+ - Renamed UnmodIterator.Implementation to UnmodIterator.Wrapper to mirror UnmodSortedIterator.Wrapper and because it's just a better name.
+ - Added tests.
+
 ## 2016-09-13 Release 2.0.9:
  - Fixed return type of ImUnsortedMap and MutableUnsortedMap.assoc(Map.Entry).  Thanks @pniederw for reporting!
  - Made ImUnsortedSet use MutableUnsortedSet's implementation of union() instead of the other way around.